Add files via upload
In this project, I have used the Streamlit web app in order to get end-to-end machine learning. There are several laptops inside the data along with their configuration and price so I have decided to go forwards with all the major steps in order to predict the price of the laptop and somehow I am able to the R2  is equal to 89%, It can be increased by using further feature engineering like hyperparameter Tunning. Thank You !!